cleanup connectivity + some bugs

+++ package/cloudshell/cp/aws/domain/services/ec2/vpc.py
@@ -2,7 +2,7 @@ from cloudshell.cp.aws.domain.services.waiters.vpc_peering import VpcPeeringConn
 class VPCService(object):
     VPC_RESERVATION = 'VPC Reservation: {0}'
 
-    def __init__(self, tag_service, subnet_service, instance_service, vpc_peering_waiter):
+    def __init__(self, tag_service, subnet_service, instance_service, vpc_waiter, vpc_peering_waiter):
         """
         :param tag_service: Tag Service
         :type tag_service: cloudshell.cp.aws.domain.services.ec2.tags.TagService
@@ -10,12 +10,15 @@ class VPCService(object):
         :type subnet_service: cloudshell.cp.aws.domain.services.ec2.subnet.SubnetService
         :param instance_service: Instance Service
         :type instance_service: cloudshell.cp.aws.domain.services.ec2.instance.InstanceService
+        :param vpc_waiter: Vpc Peering Connection Waiter
+        :type vpc_waiter: cloudshell.cp.aws.domain.services.waiters.vpc.VPCWaiter
         :param vpc_peering_waiter: Vpc Peering Connection Waiter
         :type vpc_peering_waiter: cloudshell.cp.aws.domain.services.waiters.vpc_peering.VpcPeeringConnectionWaiter
         """
         self.tag_service = tag_service
         self.subnet_service = subnet_service
         self.instance_service = instance_service
+        self.vpc_waiter = vpc_waiter
         self.vpc_peering_waiter = vpc_peering_waiter
 
     def create_vpc_for_reservation(self, ec2_session, reservation_id, cidr):
@@ -29,6 +32,8 @@ class VPCService(object):
         :return: vpc
         """
         vpc = ec2_session.create_vpc(CidrBlock=cidr)
+
+        self.vpc_waiter.wait(vpc, self.vpc_waiter.RUNNING)
 
         vpc_name = self.VPC_RESERVATION.format(reservation_id)
         self._set_tags(vpc_name=vpc_name, reservation_id=reservation_id, vpc=vpc)
@@ -84,7 +89,8 @@ class VPCService(object):
         """
         peerings = list(vpc.accepted_vpc_peering_connections.all())
         for peer in peerings:
-            peer.delete()
+            if peer.status['Code'] != 'failed':
+                peer.delete()
         return True
 
     def remove_all_security_groups(self, vpc):
@@ -107,11 +113,19 @@ class VPCService(object):
         """
         subnets = list(vpc.subnets.all())
         for subnet in subnets:
-            subnet.delete()
+            self.subnet_service.detele_subnet(subnet)
         return True
 
     def delete_all_instances(self, vpc):
         instances = list(vpc.instances.all())
-        for instance in instances:
-            self.instance_service.terminate_instance(instance)
+        self.instance_service.terminate_instances(instances)
+        return True
+
+    def delete_vpc(self, vpc):
+        """
+        Will delete the vpc instance
+        :param vpc: VPC instance
+        :return:
+        """
+        vpc.delete()
         return True
